Rename regionmanip to implicator.
